chalkpicks/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.26.0
stripe==7.10.0
apscheduler==3.10.4
python-dotenv==1.0.0
aiofiles==23.2.1
email-validator==2.1.0
